+++ spec/emojis.js
@@ -8,7 +8,7 @@ const original_timeout = jasmine.DEFAULT_TIMEOUT_INTERVAL;
 describe("Emojis", function () {
     describe("The emoji picker", function () {
 
-        beforeEach(() => (jasmine.DEFAULT_TIMEOUT_INTERVAL = 7000));
+        beforeEach(() => (jasmine.DEFAULT_TIMEOUT_INTERVAL = 10000));
         afterEach(() => (jasmine.DEFAULT_TIMEOUT_INTERVAL = original_timeout));
 
         it("can be opened by clicking a button in the chat toolbar",