+++ src/headless/converse-status.js
@@ -229,6 +229,7 @@ converse.plugins.add('converse-status', {
             if (_converse.shouldClearCache() && _converse.xmppstatus) {
                 _converse.xmppstatus.destroy();
                 delete _converse.xmppstatus;
+                api.promises.add(['statusInitialized']);
             }
         });
 
@@ -265,16 +266,19 @@ converse.plugins.add('converse-status', {
             status: {
                 /**
                  * Return the current user's availability status.
+                 * @async
                  * @method _converse.api.user.status.get
                  * @example _converse.api.user.status.get();
                  */
-                get () {
+                async get () {
+                    await api.waitUntil('statusInitialized');
                     return _converse.xmppstatus.get('status');
                 },
 
                 /**
                  * The user's status can be set to one of the following values:
                  *
+                 * @async
                  * @method _converse.api.user.status.set
                  * @param {string} value The user's chat status (e.g. 'away', 'dnd', 'offline', 'online', 'unavailable' or 'xa')
                  * @param {string} [message] A custom status message
@@ -282,7 +286,7 @@ converse.plugins.add('converse-status', {
                  * @example _converse.api.user.status.set('dnd');
                  * @example _converse.api.user.status.set('dnd', 'In a meeting');
                  */
-                set (value, message) {
+                async set (value, message) {
                     const data = {'status': value};
                     if (!Object.keys(_converse.STATUS_WEIGHTS).includes(value)) {
                         throw new Error(
@@ -292,6 +296,7 @@ converse.plugins.add('converse-status', {
                     if (isString(message)) {
                         data.status_message = message;
                     }
+                    await api.waitUntil('statusInitialized');
                     _converse.xmppstatus.save(data);
                 },
 
@@ -303,19 +308,23 @@ converse.plugins.add('converse-status', {
                  */
                 message: {
                     /**
+                     * @async
                      * @method _converse.api.user.status.message.get
                      * @returns {string} The status message
                      * @example const message = _converse.api.user.status.message.get()
                      */
-                    get () {
+                    async get () {
+                        await api.waitUntil('statusInitialized');
                         return _converse.xmppstatus.get('status_message');
                     },
                     /**
+                     * @async
                      * @method _converse.api.user.status.message.set
                      * @param {string} status The status message
                      * @example _converse.api.user.status.message.set('In a meeting');
                      */
-                    set (status) {
+                    async set (status) {
+                        await api.waitUntil('statusInitialized');
                         _converse.xmppstatus.save({ status_message: status });
                     }
                 }